# packages/contracts/openapi/arinar-v1.yaml
openapi: 3.1.0
info:
  title: Arinar API
  version: 1.0.0
  description: |
    Arinar V2 API - AI-native knowledge platform with multi-agent realtime discussions.
    
    **Model Provider Policy**: This system uses OpenRouter for LLM access, not direct OpenAI/Anthropic integrations.
  contact:
    name: Arinar Platform Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.arinar.dev
    description: Development environment
  - url: https://api.arinar.com
    description: Production
...
paths:
...
  /debates/{debate_id}/participants:
    post:
      operationId: addParticipant
      summary: Add participant to debate
      description: Adds an agent or human participant to the debate
      tags:
        - participants
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddParticipantRequest'
      responses:
        '201':
          description: Participant added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantResponse'
        '404':
          description: Debate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /workspaces/{workspace_id}/settings/models:
    get:
      operationId: getWorkspaceModels
      summary: Get workspace model defaults
      description: |
        Returns workspace-wide default models for RAG/embeddings and OCR post-processing.
        These settings sync across devices (stored in database).
        Returns system defaults if workspace hasn't configured custom values.
      tags:
        - workspace-settings
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Workspace UUID
      responses:
        '200':
          description: Workspace model defaults
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceModelsResponse'
        '403':
          description: Access denied
        '404':
          description: Workspace not found

    put:
      operationId: updateWorkspaceModels
      summary: Update workspace model defaults
      description: |
        Updates workspace-wide default models for RAG/embeddings and OCR.
        These settings:
        - Are stored server-side (sync across devices)
        - Apply automatically when requests don't specify a model
        - Do NOT require OpenRouter key (can be set before key is added)
        
        Note: OpenRouter key remains client-only (browser storage).
      tags:
        - workspace-settings
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Workspace UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceModelsRequest'
      responses:
        '200':
          description: Workspace model defaults updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceModelsResponse'
        '400':
          description: Invalid request (missing or empty model IDs)
        '403':
          description: Access denied
        '404':
          description: Workspace not found
components:
  parameters:
    DebateIdParam:
      name: debate_id
      in: path
      description: Debate session UUID
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    AddParticipantRequest:
      type: object
      required:
        - participant_type
        - role_name
      properties:
        participant_type:
          type: string
          enum: [agent, human]
        role_name:
          type: string
          minLength: 1
          maxLength: 100
        agent_config:
          type: object
          properties:
            model_id:
              type: string
              description: OpenRouter model identifier
            temperature:
              type: number
              minimum: 0
              maximum: 2
            system_prompt:
              type: string
    ParticipantResponse:
      type: object
      required:
        - participant_id
        - debate_id
        - participant_type
        - role_name
        - created_at
      properties:
        participant_id:
          type: string
          format: uuid
        debate_id:
          type: string
          format: uuid
        participant_type:
          type: string
          enum: [agent, human]
        role_name:
          type: string
        agent_config:
          type: object
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...
    WorkspaceModelsRequest:
      type: object
      required:
        - embeddings_model_id
        - ocr_model_id
      properties:
        embeddings_model_id:
          type: string
          minLength: 1
          description: "OpenRouter model ID for embeddings/RAG (e.g., moonshot/kimi-embeddings-v1)"
          example: "moonshot/kimi-embeddings-v1"
        ocr_model_id:
          type: string
          minLength: 1
          description: "OpenRouter model ID for OCR post-processing (e.g., qwen/qwen-2.5-72b-instruct)"
          example: "qwen/qwen-2.5-72b-instruct"
    WorkspaceModelsResponse:
      type: object
      required:
        - workspace_id
        - embeddings_model_id
        - ocr_model_id
        - updated_at
      properties:
        workspace_id:
          type: string
          format: uuid
          description: "Workspace UUID"
        embeddings_model_id:
          type: string
          description: "OpenRouter model ID for embeddings/RAG"
        ocr_model_id:
          type: string
          description: "OpenRouter model ID for OCR post-processing"
        updated_at:
          type: string
          format: date-time
          description: "When settings were last updated"